erp-backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-filter==25.2
python-decouple==3.8
drf-spectacular==0.29.0